add html file for current state of syntax file
TIL about the `:TOhtml()` command. This is a script that sources
2html.vim and builds the specified file into html with syntax
highlighting and line numbers. This could be useful for use in my
website as an alternative to something like hightlight.js for showing
source code. Here's hoping I don't abuse its power.

As per vim help documentation:
The ":TOhtml" command is defined in a standard plugin.
This command will source |2html.vim| for you. When a
range is given, this command sets |g:html_start_line|
and |g:html_end_line| to the start and end of the
range, respectively. Default range is the entire
buffer.

If the current window is part of a |diff|, unless
|g:html_diff_one_file| is set, :TOhtml will convert
all windows which are part of the diff in the current
tab and place them side-by-side in a <table> element
in the generated HTML. With |g:html_line_ids| you can
jump to lines in specific windows with (for example)

Examples: >

	:10,40TOhtml " convert lines 10-40 to html
	:'<,'>TOhtml " convert current/last visual selection
	:TOhtml      " convert entire buffer